Replaced MainBus with ErrorsBus and TimerBus.
Cleared out some of the console.logs related to API events as I'm only interested in errors and those are now centralized in the ErrorsBus.

Fixed a bug in Processing.vue where promise .then is triggere if all promises are caught beforehand.

Shufled around contents of main to make it more readable.