3d attempt at fixing duplicate entries issue. I think this will fix it.
Moved TimerBus listeners under Project to the parent Project component.

After a few hours of troubleshooting, I finally found the root cause of the issue. I am currently using a separate Vue instance to function as an event bus. Th reason why I am seeing multiple entries is that when a Vue instance is created (this part I don’t fully understand yet), event listeners ( vm.$emit and vm.$on) are registered in the Vue instance.

When a new project is created, it adds its own event listeners to the event bus. Meaning the event bus now has multiple event listeners registered under the same name. When the event listener is triggered, the events are fired multiple times. This also occurs for many things, including the issue I noted with the complete entry.